requirements.txt
python-whois == 0.8.0
requests-html == 0.10.0